Refactor SaveNXcanSAS algorithm to support different variants
This is a squashed version of #38582

Remove helper and main exec structure out of inherited saveNXCanSass, inherit from saveNXcanSASBase

Add parent SaveNXcanSasBase class with helper functionsin anonymous namespace

Add file for NXcanSAS helper functions

Referene to helper file in tests

Organize code, add some descriptions and refactor some individual lines

Fix typo in include path on cmakelist

Fix cppcheck style warning:

Fix doxygen documentation warnings:
1. There were doxygen lines in add2DData function that needed to be escaped with \. to avoid a warning, this was not detected before
as doxygen does not look in anonymous namespaces for mantid.
2. There was a doxygen comment missing input parameters to the function addInstrument.

Update header date on new file
